DataCo SMART SUPPLY CHAIN FOR BIG DATA
Collected this dataset from Kaggle. Checked for the incorrect dtypes and changed it. Data 
Imbalance is checked and found out it is balanced dataset. Drawn insights of the data with the help of 
seaborn and matplotlib plots. Done Statistical tests to find the features significance to the target variable. 
Encoded the Categorical variables and the data is split into train and test set. Built the base model as logistic 
regression and advance model as Random Forest Classifier.