Host Price Prediction
Required data cleaning were done as a part of EDA. Association of features with target variables 
were studied using visualization graphs and asserted with the statistical test. This involved a detailed 
analysis on the data and data pre-processing. Feature engineering based on wordcloud is done to obtain 
more attributes. Random Forest models are built for the optimal calculation of prices.