Predicting Air Quality
Raw data acquired from kaggle. Cleaned the data which has impurities and impute the missing values in 
the data using KNN-Imputer. Performed statistical test to find the significance of the variable with the target. Finally 
build the various ML model such as linear and non-linear models from the we find Xgboost Regressor as the best 
model for this data which has minimum error comparing to other models